refactor(writer): remove legacy writing models
