+++ lazyllm/tools/writer/data_models/__init__.py
@@ -8,17 +8,7 @@ from .context import (
     StyleProfile,
     WritingContext,
 )
-from .writing import (
-    DraftBlock,
-    DraftDocument,
-    DraftSection,
-    OutlineNode,
-    SectionInstruction,
-    SectionInstructionList,
-    WritingOutline,
-    WritingOutput,
-    WritingSubTask,
-)
+from .planning import SectionInstruction, SectionInstructionList
 from .revision import Anchor, LocateResult, ModifyInstruction, ModifyPlan, PatchBlock, PatchHunk, PatchResult, PatchSet
 from .multimodal import MediaAsset, MediaAssetLibrary, VisualInstruction
 from .quality import AuditIssue, AuditResult, ReviewReport
@@ -28,7 +18,6 @@ __all__ = [
     'InputResource',
     'Selection',
     'TargetDocument',
-    'WritingOutput',
     'WritingTask',
     'MaterialStyle',
     'ResourceProfile',
@@ -38,14 +27,8 @@ __all__ = [
     'DocumentSummary',
     'StyleProfile',
     'WritingContext',
-    'DraftBlock',
-    'DraftDocument',
-    'DraftSection',
-    'OutlineNode',
     'SectionInstruction',
     'SectionInstructionList',
-    'WritingOutline',
-    'WritingSubTask',
     'Anchor',
     'ModifyInstruction',
     'ModifyPlan',
